feat: Refactor game structure by moving classes to models and updating imports
