+++ models/RiddleManager.js
