+++ models/UIManager.js
